requirements.txt
markitdown[all]
streamlit
openai
tiktoken
pytesseract
PyMuPDF
Pillow
